add purchase and some details in personal center

+++ Art_Sea/php/删购.php
@@ -3,13 +3,15 @@ include("functions.php");
 
 session_start();
 
-$connect = connectDB();
-$artworkID = $_GET['artworkID'];
-$userID = $_SESSION['userID'];
-$sql = "DELETE FROM carts WHERE userID='" . $userID . "' AND artworkID='" . $artworkID . "'";
-if ($connect->query($sql)) {
-    echo 'delete from cart successfully';
-}else{
-    echo "deleting error:" . $connect->error;
+if (isset($_SESSION['admin']) && $_SESSION['admin'] === TRUE) {
+    $connect = connectDB();
+    $artworkID = $_GET['artworkID'];
+    $userID = $_SESSION['userID'];
+    $sql = "DELETE FROM carts WHERE userID='" . $userID . "' AND artworkID='" . $artworkID . "'";
+    if ($connect->query($sql)) {
+        echo 'delete from cart successfully';
+    } else {
+        echo "deleting error:" . $connect->error;
+    }
 }
 ?>
